End Section 18

+++ client/src/app/messages/messages.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Message } from '../_models/message';
 import { Pagination } from '../_models/pagination';
+import { ConfirmService } from '../_services/confirm.service';
 import { MessageService } from '../_services/message.service';
 
 @Component({
@@ -17,7 +18,7 @@ export class MessagesComponent implements OnInit {
   pageSize: number = 5;
   loading: boolean = false;
 
-  constructor(private messageService: MessageService) { }
+  constructor(private messageService: MessageService, private confirmService: ConfirmService) { }
 
   ngOnInit(): void {
     this.loadMessages();
@@ -36,11 +37,15 @@ export class MessagesComponent implements OnInit {
   }
 
   deleteMessage(id) {
-    this.messageService
-      .deleteMessage(id)
-      .subscribe(() => {
-        this.messages.splice(this.messages.findIndex(m => m.id === id), 1);
-      })
+    this.confirmService.confirm("Delete Message", "Message will be deleted. This cannot be undone.", "Delete").subscribe(result => {
+      if (result) {
+        this.messageService
+          .deleteMessage(id)
+          .subscribe(() => {
+            this.messages.splice(this.messages.findIndex(m => m.id === id), 1);
+          })
+      }
+    });
   }
 
   pageChanged(event: any) {
